feat: Add @radix-ui/react-accordion dependency and update PlacesSection
The commit adds the "@radix-ui/react-accordion" dependency to the package.json file. This dependency will be used to implement an accordion component in the application.

In addition, the PlacesSection component is updated to accept a sectionTitle and sectionDescription prop. This allows for more flexibility in customizing the title and description of the section.

These changes enhance the functionality and user experience of the application.